% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlhelpers-package.R
\docType{package}
\name{htmlhelpers-package}
\alias{htmlhelpers}
\alias{htmlhelpers-package}
\title{htmlhelpers: Helper Functions for Those Who Can't Write HTML Code for Media Files}
\description{
The `htmlhelpers` package is for those of us who understand what an html file is, and what it is for, but cannot write html code by heart. The package makes it easy to embed media files (videos, images, PDFs) into your Quarto documents, R Markdown files, and Shiny apps without writing HTML code manually.
}
\author{
\strong{Maintainer}: Dan Swart \email{danswartcpa@gmail.com}

}
\keyword{internal}
